Separate get account transactions and get account asset transactions
