+++ shared_model/interfaces/queries/get_account_transactions.hpp
@@ -19,48 +19,11 @@
 #include "interfaces/primitive.hpp"
 #include "model/queries/get_transactions.hpp"
 
-#ifndef IROHA_SHARED_MODEL_GET_TRANSACTIONS_HPP
-#define IROHA_SHARED_MODEL_GET_TRANSACTIONS_HPP
+#ifndef IROHA_SHARED_MODEL_GET_ACCOUNT_TRANSACTIONS_HPP
+#define IROHA_SHARED_MODEL_GET_ACCOUNT_TRANSACTIONS_HPP
 
 namespace shared_model {
   namespace interface {
-
-    /**
-     * Query for getting transactions of given asset of an account
-     */
-    class GetAccountAssetTransactions
-        : public Primitive<GetAccountAssetTransactions,
-                           iroha::model::GetAccountAssetTransactions> {
-     public:
-      /**
-       * @return account_id of requested transactions
-       */
-      virtual const types::AccountIdType &accountId() const = 0;
-      /**
-       * @return assetId of requested transactions
-       */
-      virtual const types::AccountIdType &assetId() const = 0;
-
-      OldModelType *makeOldModel() const override {
-        auto oldModel = new iroha::model::GetAccountAssetTransactions;
-        oldModel->account_id = accountId();
-        oldModel->asset_id = assetId();
-        return oldModel;
-      }
-
-      std::string toString() const override {
-        return detail::PrettyStringBuilder()
-            .init("GetAccountAssetTransactions")
-            .append("account_id", accountId())
-            .append("asset_id", assetId())
-            .finalize();
-      }
-
-      bool operator==(const ModelType &rhs) const override {
-        return accountId() == rhs.accountId() and assetId() == rhs.assetId();
-      }
-    };
-
     /**
      * Query for getting transactions of account
      */
@@ -93,4 +56,4 @@ namespace shared_model {
   }  // namespace interface
 }  // namespace shared_model
 
-#endif  // IROHA_SHARED_MODEL_GET_TRANSACTIONS_HPP
+#endif  // IROHA_SHARED_MODEL_GET_ACCOUNT_TRANSACTIONS_HPP
