+++ shared_model/interfaces/queries/query.hpp
@@ -25,10 +25,11 @@
 #include "interfaces/primitive.hpp"
 #include "interfaces/queries/get_account.hpp"
 #include "interfaces/queries/get_account_assets.hpp"
+#include "interfaces/queries/get_account_asset_transactions.hpp"
+#include "interfaces/queries/get_account_transactions.hpp"
 #include "interfaces/queries/get_asset_info.hpp"
 #include "interfaces/queries/get_roles.hpp"
 #include "interfaces/queries/get_signatories.hpp"
-#include "interfaces/queries/get_transactions.hpp"
 #include "interfaces/signable.hpp"
 #include "interfaces/visitor_apply_for_all.hpp"
 #include "model/query.hpp"
